Split ping-pong
