+++ examples-other/simgrid/master-workers/fwd.h
@@ -1,0 +1,5 @@
+#pragma once
+
+namespace simgrid::s4u {
+class Mailbox;
+}
