+++ examples-other/simgrid/ping-pong/CMakeLists.txt
@@ -1,4 +1,7 @@
-add_executable(ping-pong ping-pong.cpp)
+add_executable(ping-pong
+    ping-pong.cpp
+    process.cpp
+)
 
 target_link_libraries(ping-pong ${SimGrid_LIBRARY})
 target_include_directories(ping-pong PUBLIC SYSTEM ${SimGrid_INCLUDE_DIR})
